requirements.txt
requests>=2.31.0
playwright>=1.39.0
python-dotenv>=1.0.0
